 requirements.txt
Flask==2.3.3
pymysql==1.1.0
python-dotenv==1.0.0
gunicorn==21.2.0